# docs/swagger.yaml
definitions:
  main.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      price:
        type: number
      title:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      description: Возвращает список всех товаров
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список товаров
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Добавляет новый товар в базу данных
      parameters:
      - description: Данные товара
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить товар
      tags:
      - products
  /products/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет данные товара по его ID
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные товара
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить товар
      tags:
      - products
  /products/{id}/history:
    get:
      description: Возвращает историю изменений товара по его ID
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить историю товара
      tags:
      - products
  /products/{id}/rollback:
    post:
      description: Откатывает товар к указанной версии
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      - description: Версия для отката
        in: query
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Откатить товар
      tags:
      - products
swagger: "2.0"