# swagger.go
package main

import (
	"database/sql"
	"net/http"
)

// @title Product API
// @version 1.0
// @description API для управления товарами
// @host localhost:8080
// @BasePath /api

// Product представляет модель товара.
type Product struct {
	ID          int     `json:"id"`
	Title       string  `json:"title"`
	Description *string `json:"description,omitempty"`
	Price       float64 `json:"price"`
	Version     int     `json:"version"`
}

// GetProducts возвращает список всех товаров.
// @Summary Получить список товаров
// @Description Возвращает список всех товаров
// @Tags products
// @Produce json
// @Success 200 {array} Product
// @Failure 500 {object} map[string]string
// @Router /products [get]
func GetProducts(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Реализация функции
	}
}

// AddProduct добавляет новый товар.
// @Summary Добавить товар
// @Description Добавляет новый товар в базу данных
// @Tags products
// @Accept json
// @Produce json
// @Param product body Product true "Данные товара"
// @Success 201 {object} Product
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /products [post]
func AddProduct(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Реализация функции
	}
}

// UpdateProduct обновляет существующий товар.
// @Summary Обновить товар
// @Description Обновляет данные товара по его ID
// @Tags products
// @Accept json
// @Produce json
// @Param id path int true "ID товара"
// @Param product body Product true "Новые данные товара"
// @Success 200 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /products/{id} [put]
func UpdateProduct(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Реализация функции
	}
}

// RollbackProduct откатывает товар к предыдущей версии.
// @Summary Откатить товар
// @Description Откатывает товар к указанной версии
// @Tags products
// @Produce json
// @Param id path int true "ID товара"
// @Param version query int true "Версия для отката"
// @Success 200 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /products/{id}/rollback [post]
func RollbackProduct(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Реализация функции
	}
}

// GetProductHistory возвращает историю изменений товара.
// @Summary Получить историю товара
// @Description Возвращает историю изменений товара по его ID
// @Tags products
// @Produce json
// @Param id path int true "ID товара"
// @Success 200 {array} map[string]interface{}
// @Failure 500 {object} map[string]string
// @Router /products/{id}/history [get]
func GetProductHistory(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Реализация функции
	}
}